+++ public/errors/consent.php
@@ -1,0 +1,5 @@
+<?php
+
+$config = SimpleSAML\Configuration::getInstance();
+$t = new SimpleSAML\XHTML\Template($config, 'authoauth2:errors/consent.twig');
+$t->send();